change int to double in index in polynomial data structure

+++ NumericalCalulator/Common/polynomial.h
@@ -15,7 +15,7 @@ typedef struct link
     double Num;
 	struct link *Next;
     char Operator;
-	int Index;
+	double Index;
 }Link;
 
  class LinkList
